jarvis_project/requirements.txt
Django==5.2.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.1
dj-database-url==2.2.0
psycopg2-binary==2.9.9
openai==1.51.2
django-cors-headers==4.4.0
gunicorn==22.0.0
